+++ include/openpose/pose/wPoseExtractor.hpp
@@ -87,11 +87,11 @@ namespace op
                     // Keep desired top N people
                     spPoseExtractor->keepTopPeople(tDatum.poseKeypoints, tDatum.poseScores);
                     // ID extractor (experimental)
-                    tDatum.poseIds = spPoseExtractor->extractIdsLockThread(tDatum.poseKeypoints, tDatum.cvInputData,
-                                                                           i, tDatum.id);
+                    tDatum.poseIds = spPoseExtractor->extractIdsLockThread(
+                        tDatum.poseKeypoints, tDatum.cvInputData, i, tDatum.id);
                     // Tracking (experimental)
-                    spPoseExtractor->trackLockThread(tDatum.poseKeypoints, tDatum.poseIds, tDatum.cvInputData, i,
-                                                     tDatum.id);
+                    spPoseExtractor->trackLockThread(
+                        tDatum.poseKeypoints, tDatum.poseIds, tDatum.cvInputData, i, tDatum.id);
                 }
                 // Profiling speed
                 Profiler::timerEnd(profilerKey);