+++ doc/release_notes.md
@@ -288,6 +288,7 @@ OpenPose Library - Release Notes
     17. Improved and added several functions to `utilities/keypoints.hpp` and Array to simplify keypoint post-processing.
     18. Removed warnings from Spinnaker SDK at compiling time.
     19. All bash scripts incorporate `#!/bin/bash` to tell the terminal that they are bash scripts.
+    20. Added flag `--verbose` to plot the progress.
 2. Functions or parameters renamed:
     1. By default, python example `tutorial_developer/python_2_pose_from_heatmaps.py` was using 2 scales starting at -1x736, changed to 1 scale at -1x368.
     2. WrapperStructPose default parameters changed to match those of the OpenPose demo binary.